Improve information for logging and for storage
The version set for the store is the version of the project.
The comment is logging the user and th amchine uploaded symbols.
Verbose symstore logging is printed when debug level is set for Gradle.
Make sure symstore logs are printed even when something went wrong.

Other stuff:
Parameters always quoted